Feature: 훅과 컴포넌트 분리
