+++ src/providers/index.tsx
@@ -1,7 +1,3 @@
-export { LoginProvider, useLoginContext, LoginContext } from "./LoginProvider";
-export {
-  NotificationProvider,
-  useNotificationContext,
-  NotificationContext,
-} from "./NotificationProvider";
-export { ThemeProvider, useThemeContext, ThemeContext } from "./ThemeProvider";
+export { LoginProvider } from "./LoginProvider";
+export { NotificationProvider } from "./NotificationProvider";
+export { ThemeProvider } from "./ThemeProvider";
